Refactoring add composable for filtering data
