+++ src/modules/product/components/ProductList.vue
@@ -1,11 +1,12 @@
 <script lang="ts" setup>
 import { ref, computed, onBeforeMount } from 'vue'
-import useProductData from '@product/composables/useProductListData'
 import { useCartStore } from '@cart/stores/cart'
 import type ProductType from '@product/types/ProductType'
 import ProductFilters from '@product/components/ProductFilters.vue'
 
-const { loading, productList, getData } = useProductData()
+import useFilters from '@product/composables/useFiltersProduct'
+
+const { loading, applyFilters, filteredProducts } = useFilters()
 
 const cartStore = useCartStore()
 const appliedFilters = ref<{ key: keyof ProductType; value: string | number; type: 'string' | 'number' }[]>([])
@@ -16,7 +17,7 @@ const limitedDescription = (description: string) => {
 }
 
 onBeforeMount(async () => {
-  await getData()
+  await applyFilters()
 })
 
 const viewProduct = (id: number) => {
@@ -27,35 +28,6 @@ const addToCart = (product: ProductType) => {
   cartStore.addToCart(product)
   console.log('Add to Cart', product)
 }
-
-const applyFilters = async (filters: { key: keyof ProductType; value: string | number; type: 'string' | 'number' }[]) => {
-  console.log("Filters", filters);
-  appliedFilters.value = filters
-
-  const filterApiArray = [] as { key: string; value: string | number }[]
-  filters.forEach(filter => {
-    if (filter.key === 'category') {
-      filterApiArray.push({ key: filter.key, value: filter.value.toString() });
-    }
-  });
-
-    await getData(filterApiArray)
-}
-
-
-const filteredProducts = computed(() => {
-  return productList.value.filter((product: ProductType) => {
-    return appliedFilters.value.every(filter => {
-      if (filter.key === 'category') return true
-      const productValue = filter.key.split('.').reduce((o, i) => o[i], product)
-      if (filter.type === 'number') {
-        return productValue >= filter.value
-      } else {
-        return productValue.toString().toLowerCase().includes(filter.value.toString().toLowerCase())
-      }
-    })
-  })
-})
 </script>
 
 <template>
@@ -64,7 +36,7 @@ const filteredProducts = computed(() => {
       <v-card-title>
         <v-row align="center">
           <v-col cols="4">
-            <h3>Product List ({{ productList.length }})</h3>
+            <h3>Product List ({{ filteredProducts.length }})</h3>
           </v-col>
           <v-col cols="8" class="d-flex justify-end">
             <ProductFilters @filter="applyFilters" />
